feat(update): Move implemented attributes and qualifiers to top level
