+++ composify/metadata/__init__.py
@@ -1,16 +1,6 @@
-from .attributes import (
-    AttributeSet,
-    BaseAttributeMetadata,
-    Name,
-    collect_attributes,
-)
+from .attributes import AttributeSet, BaseAttributeMetadata, collect_attributes
 from .base import BaseMetadata, MetadataSet, collect_metadata
-from .qualifiers import (
-    BaseQualifierMetadata,
-    DisallowSubclass,
-    QualifierSet,
-    collect_qualifiers,
-)
+from .qualifiers import BaseQualifierMetadata, QualifierSet, collect_qualifiers
 
 __all__ = [
     "BaseMetadata",
