+++ composify/metadata/attributes.py
@@ -1,8 +1,6 @@
-from dataclasses import dataclass
-
 from composify.types import AnnotatedType
 
-from .base import SLOTS, BaseMetadata, MetadataSet, _collect_metadata
+from .base import BaseMetadata, MetadataSet, _collect_metadata
 
 
 class BaseAttributeMetadata(BaseMetadata):
@@ -16,15 +14,3 @@ class AttributeSet(MetadataSet[BaseAttributeMetadata]):
 def collect_attributes(type_: AnnotatedType) -> AttributeSet:
     """Collect all annotated metadata that inherits BaseAttributeMetadata class as a frozenset."""
     return _collect_metadata(type_, BaseAttributeMetadata, AttributeSet)
-
-
-class Name(str, BaseAttributeMetadata):
-    def __repr__(self) -> str:
-        return f"Name({self})"
-
-
-@dataclass(frozen=True, **SLOTS)
-class ProvidedBy(BaseAttributeMetadata):
-    """Describe the provider of the value."""
-
-    source: str
